requirement.txt
asgiref==3.2.7
crispy-forms-materialize==0.2
Django==3.0.5
django-crispy-forms==1.9.0
django-environ==0.4.5
django-materialize-css==0.0.1
djangorestframework==3.11.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2==2.8.5
pytz==2019.3
sqlparse==0.3.1
